Split probing and visualization into two programs
because their parameterizations are almost completely independent.
