+++ examples/on-stability-testing.md
@@ -91,12 +91,15 @@ The analysis phase is carried out by the program
 - The program is configured by editing global variables in its
   source code.
 
-The probing and visualization phases are carried out by the program
+The probing phase is carried out by the program
 `examples/satellites/check_stability.py`
 - The probe set and configuration is defined in the source of the
   program.
-- Probing can be skipped by commenting it out.
-- Visualization is determined entirely by the probe results.
+
+The visualization phase is carried out by the program
+`examples/satellites/visualize.py`
+- The configuration is defined in the source of the program.
+- Visualization content is determined entirely by the probe results.
 
 Generic probing and visualization machinery is available in the
 modules `bdbcontrib.experiments.probe.py` and
