+++ examples/satellites/check_stability.py
@@ -14,14 +14,9 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import matplotlib
-matplotlib.use("Agg")
-
 import math
-import re
 
 from bdbcontrib.experiments.probe import log, probe_fileset
-from bdbcontrib.experiments.visualization import plot_results
 
 ######################################################################
 ## Probes                                                          ##
@@ -171,16 +166,4 @@ def save_probe_results(filename):
         pickle.dump(results, f)
     log("Saved probe results to %s" % filename)
 
-def analysis_count_from_file_name(fname):
-    return int(re.match(r'.*-[0-9]*m-([0-9]*)i.bdb', fname).group(1))
-
-def plot_probe_results(filename):
-    log("Loading probe results from %s" % filename)
-    with open(filename, "r") as f:
-        results = pickle.load(f)
-    results = [((probe, n_models, analysis_count_from_file_name(fname)), val)
-               for ((fname, n_models, probe), val) in results.iteritems()]
-    plot_results(results, outdir="figures-3200m-30i")
-
 save_probe_results("results-3200m-30i.pkl")
-plot_probe_results("results-3200m-30i.pkl")
